+++ app/storage/upload-file-to-blob.js
@@ -1,4 +1,4 @@
-const { blobServiceClient } = require('../blob-storage')
+const { blobServiceClient } = require('./blob-storage')
 const { v4: uuidv4 } = require('uuid')
 const storageConfig = require('../config/storage')
